test: migrate integrations-github-label-listeners to bdd
